requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
geopandas==1.1.1
idna==3.11
internetarchive==5.5.1
jsonpatch==1.33
jsonpointer==3.1.1
numpy==1.26.4
packaging==26.0
pandas==2.2.2
pillow==12.2.0
pyarrow==20.0.0
pyogrio==0.11.0
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.33.1
setuptools==70.0.0
shapely==2.1.2
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.2
xmltodict==0.13.0
